Restore modern Arch build compatibility

+++ src/include/game/game_data.h
@@ -71,7 +71,7 @@ typedef struct Data {
 typedef struct AI {
   int active;
   int tdiff;
-  long lasttime;
+  unsigned int lasttime;
 	segment2 left, right, front, backleft;
 } AI;
